ai-task-automation-agent/backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
pydantic-settings>=2.0.0
groq>=0.4.0
httpx>=0.26.0
beautifulsoup4>=4.12.0
python-multipart>=0.0.6
aiofiles>=23.0.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
alembic>=1.13.0
apscheduler>=3.10.0
slowapi>=0.1.9
google-api-python-client>=2.116.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
